split layer examples
